added boilerplate code and renamed oldScraper.py
